modify simple broadcaster and listener example
